fix: change reward amount to u32

+++ contracts/appchain/src/models/reward.cairo
@@ -7,6 +7,6 @@ pub struct Reward {
     pub player: ContractAddress,
     #[key]
     pub game_id: u32,
-    pub amount: u16,
+    pub amount: u32,
     pub claimed: bool,
 }
